requirements.txt
flask
datetime
netifaces
requests
psutil
hurry.filesize==0.9
GPUtil
multiprocess
